+++ src/seabreeze/pyseabreeze/devices.py
@@ -402,6 +402,7 @@ class USB2000PLUS(SeaBreezeDevice):
     feature_classes = (
         sbf.eeprom.SeaBreezeEEPromFeatureOOI,
         sbf.spectrometer.SeaBreezeSpectrometerFeatureUSB2000PLUS,
+        sbf.continuousstrobe.SeaBreezeContinuousStrobeFeatureOOI,
         sbf.rawusb.SeaBreezeRawUSBBusAccessFeature,
     )